+++ test/test_sphere.jl
@@ -1,60 +1,9 @@
 using LinearAlgebra
 using ManifoldsBase
-using ManifoldsBase: ℝ, ℂ, DefaultManifold
+using ManifoldsBase: ℝ, ℂ, DefaultManifold, RealNumbers
 using Test
 
-# minimal sphere implementation for testing more complicated manifolds
-struct TestSphere{N,𝔽} <: AbstractManifold{𝔽} end
-TestSphere(N::Int, 𝔽 = ℝ) = TestSphere{N,𝔽}()
-
-ManifoldsBase.representation_size(::TestSphere{N}) where {N} = (N + 1,)
-
-function ManifoldsBase.exp!(M::TestSphere, q, p, X)
-    return exp!(M, q, p, X, one(number_eltype(X)))
-end
-function ManifoldsBase.exp!(::TestSphere, q, p, X, t::Number)
-    θ = norm(X)
-    if θ == 0
-        copyto!(q, p)
-    else
-        X_scale = t * sin(θ) / θ
-        q .= p .* cos(θ) .+ X .* X_scale
-    end
-    return q
-end
-
-ManifoldsBase.inner(::TestSphere, p, X, Y) = dot(X, Y)
-
-function ManifoldsBase.inverse_retract_project!(M::TestSphere, X, p, q)
-    X .= q .- p
-    project!(M, X, p, X)
-    return X
-end
-
-function ManifoldsBase.log!(::TestSphere, X, p, q)
-    cosθ = clamp(real(dot(p, q)), -1, 1)
-    X .= q .- p .* cosθ
-    nrmX = norm(X)
-    if nrmX > 0
-        X .*= acos(cosθ) / nrmX
-    end
-    return X
-end
-
-function ManifoldsBase.project!(::TestSphere, q, p)
-    q .= p ./ norm(p)
-    return q
-end
-
-function ManifoldsBase.project!(::TestSphere, Y, p, X)
-    Y .= X .- p .* real(dot(p, X))
-    return Y
-end
-
-# from Absil, Mahony, Trumpf, 2013 https://sites.uclouvain.be/absil/2013-01/Weingarten_07PA_techrep.pdf
-function ManifoldsBase.Weingarten!(::TestSphere, Y, p, X, V)
-    return Y .= -X * p' * V
-end
+include("test_manifolds.jl")
 
 @testset "TestSphere" begin
     @testset "ShootingInverseRetraction" begin